requirements.txt
flask==3.1.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
anthropic==0.42.0
